refactor: move Usage Profile view methods to view.py
